% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessment.R
\name{compare_Theta}
\alias{compare_Theta}
\title{Compute the differences between two Mutual Hazard Networks}
\usage{
compare_Theta(true_Theta, pred_Theta, q = 0.01)
}
\arguments{
\item{true_Theta}{A ground truth MHN represented by a square matrix.}

\item{pred_Theta}{An estimated MHN represented by a square matrix.}

\item{q}{A threshold to zero out very small entries in the estimated MHN (Default: 1e-2).}
}
\value{
A vector of performance metrics:
\itemize{
\item SHD: Structural Hamming Distance between the two matrices;
\item TP: True positives in the estimated MHN;
\item FP: False positives in the estimated MHN;
\item TN: True negatives in the estimated MHN;
\item FN: False negatives in the estimated MHN;
\item Precision: True positives divided by the total number of edges in the estimated MHN;
\item TPR (Recall): True positives divided by the total number of edges in the true MHN;
\item FPR_N: False positives divided by the total number of non-edges in the true MHN;
\item FPR_P: False positives divided by the total number of edges in the true MHN;
\item MSE: Mean squared error between the two matrices.
}
}
\description{
This function computes the differences between two Mutual Hazard Networks.
}
\author{
Xiang Ge Luo
}
