% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{generate_trees}
\alias{generate_trees}
\title{generate_trees(n):
a function that generates mutation trees according an MHN}
\usage{
generate_trees(
  n = 10,
  N = 1,
  lambda_s = 1,
  Theta = NULL,
  sparsity = 0.5,
  mutations = NULL
)
}
\arguments{
\item{n:}{number of mutations being considered}

\item{N:}{number of trees being generated (default: 1)}

\item{model_par:}{list of model parameters}
}
\value{
a list of trees
}
\description{
generate_trees(n):
a function that generates mutation trees according an MHN
}
