% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessment.R
\name{get_hintra_pathways}
\alias{get_hintra_pathways}
\title{Compute HINTRA pathway probabilities}
\usage{
get_hintra_pathways(n, tree_df, n_order = 4)
}
\arguments{
\item{n}{Number of mutational events.}

\item{tree_df}{A data frame with the following columns:
\itemize{
\item Tree_ID: IDs of mutation trees, unique for each patient
\item Node_ID: IDs of each node in the tree, including the root node (with ID "1"), unique for each node
\item Mutation_ID: IDs of each mutational event, the root node has a mutation ID of "0",
other mutation IDs can be duplicated in the tree to allow for parallel mutations
\item Parent_ID: IDs of the parent node ID. The root node has itself as parent (ID "1").
}}

\item{n_order}{Length of the pathways (Default: 4).}
}
\value{
Pathway probabilities inferred from the row-normalized beta matrix of the HINTRA algorithm.
}
\description{
This function computes the pathway probabilities inferred from the
row-normalized beta matrix of the HINTRA algorithm.
}
\author{
Xiang Ge Luo
}
