% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn_MHN.R
\name{get_timed_trees}
\alias{get_timed_trees}
\title{get_timed_trees(n, N, trees, Theta, lambda_s, M, MC_flags)
A function that computes the expected waiting times of all edges in a set of trees
given a mutual hazard network}
\usage{
get_timed_trees(n, N, trees, Theta, lambda_s, M, MC_flags)
}
\arguments{
\item{n}{Number of events in the MHNs}

\item{N}{Sample size}

\item{trees}{Mutation tree structures}

\item{Theta}{An n-by-n matrix representing the mutual hazard network}

\item{lambda_s}{Rate of the sampling event}

\item{M}{Number of Monte Carlo copies}

\item{MC_flags}{A boolean vector indicating whether Monte Carlo sampling is used for each tree}
}
\value{
Timed trees
}
\description{
get_timed_trees(n, N, trees, Theta, lambda_s, M, MC_flags)
A function that computes the expected waiting times of all edges in a set of trees
given a mutual hazard network
}
