% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn_MHN.R
\name{initialize_Theta}
\alias{initialize_Theta}
\title{initialize_Theta(n, N, trees, lambda_s)
A function that initializes the diagonal entries of a MHN based on the
relative frequencies of each event starting from root.}
\usage{
initialize_Theta(n, N, trees, lambda_s)
}
\arguments{
\item{n}{Number of events in the MHNs}

\item{N}{Sample size}

\item{trees}{Mutation tree structures}

\item{lambda_s}{Rate of the sampling event}
}
\value{
An n-by-n matrix representing the initialized MHN
}
\description{
initialize_Theta(n, N, trees, lambda_s)
A function that initializes the diagonal entries of a MHN based on the
relative frequencies of each event starting from root.
}
