% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_output.R
\name{input_tree_df}
\alias{input_tree_df}
\title{Process a dataframe of mutation trees and output a TreeMHN object}
\usage{
input_tree_df(
  n,
  tree_df,
  patients = NULL,
  tree_labels = NULL,
  mutations = NULL,
  weights = NULL
)
}
\arguments{
\item{n}{Number of mutational events.}

\item{tree_df}{A data frame with the following columns:
\itemize{
\item Patient_ID: IDs of patients, unique for each patient
\item Tree_ID: IDs of mutation trees, unique for each tree
\item Node_ID: IDs of each node in the tree, including the root node (with ID "1"), unique for each node
\item Mutation_ID: IDs of each mutational event, the root node has a mutation ID of "0",
other mutation IDs can be duplicated in the tree to allow for parallel mutations
\item Parent_ID: IDs of the parent node ID. The root node has itself as parent (ID "1").
}}

\item{patients}{A list of patient labels, which must be unique values.
If no labels are given, then the patient IDs will be used.}

\item{tree_labels}{A list of tree labels, which must be unique values.
If no labels are given, then the tree IDs will be used.}

\item{mutations}{A list of mutation names, which must be unique values.
If no names are given, then the mutation IDs will be used.}

\item{weights}{Weights of the trees. If no values are given, weights are
assigned equally to the trees such that each patient has a weight of 1.}
}
\value{
A TreeMHN object
}
\description{
This function processes a dataframe of mutation trees and output a TreeMHN object.
}
\author{
Xiang Ge Luo
}
