% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessment.R
\name{KL_divergence}
\alias{KL_divergence}
\title{Compute the KL divergence between two probability distributions}
\usage{
KL_divergence(p, q)
}
\arguments{
\item{p}{A probability distribution.}

\item{q}{Another probability distribution.}
}
\value{
KL(p || q)
}
\description{
This function computes the KL divergence between two probability distributions.
}
