% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn_MHN.R
\name{learn_MHN}
\alias{learn_MHN}
\title{Learn a Mutual Hazard Network from a set of mutation trees}
\usage{
learn_MHN(
  tree_obj,
  gamma = 0.5,
  lambda_s = 1,
  Theta_init = NULL,
  M = 100,
  iterations = 1000,
  to_mask = integer(0),
  use_EM = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{tree_obj}{A TreeMHN object}

\item{gamma}{Penalization parameter in the objective function (Default: 0.5)}

\item{lambda_s}{Sampling rate (Default: 1)}

\item{Theta_init}{Initial value of the MHN provided to the optimization procedure (Default: NULL)}

\item{M}{Number of Monte Carlo samples to be drawn (Default: 100)}

\item{iterations}{Number of iterations for the EM/MCEM algorithm (Default: 1000)}

\item{to_mask}{An integer vector of indices by column, which is used to mask the
off-diagonal entries of an MHN (Default: an empty vector)}

\item{use_EM}{A boolean value to determine whether the EM/MCEM algorithm is used (Default: FALSE)}

\item{verbose}{A boolean value to determine whether optimization steps are printed (Default: FALSE)}
}
\value{
A Mutual Hazard Network Theta
}
\description{
This function learns a Mutual Hazard Network from a set of mutation trees
in the format of a TreeMHN object
}
\author{
Xiang Ge Luo
}
