% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fns.R
\name{next_mutation}
\alias{next_mutation}
\title{Plot the next most probable mutational events}
\usage{
next_mutation(n, tree, Theta, mutations = NULL, tree_label = NULL, top_M = 1)
}
\arguments{
\item{n}{Number of mutational events.}

\item{tree}{A tree in named list format.}

\item{mutations}{A list of mutation names, which must be unique values.
If no names are given, then the mutation IDs will be used.}

\item{tree_label}{The title of the tree (Default: NULL).}

\item{top_M}{Number of most probable mutational events to plot (Default: 1).}
}
\description{
Given a particular tree and a Mutual Hazard Network, this function 
finds the next most probable mutational events.
}
