% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_output.R
\name{output_tree_df}
\alias{output_tree_df}
\title{output_tree_df(tree_obj)
A helper function that construct a tree data frame from a list of mutation trees}
\usage{
output_tree_df(tree_obj)
}
\arguments{
\item{tree_obj}{A TreeMHN object}
}
\value{
A TreeMHN object with the tree data frame
}
\description{
output_tree_df(tree_obj)
A helper function that construct a tree data frame from a list of mutation trees
}
