% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fns.R
\name{plot_next_mutations}
\alias{plot_next_mutations}
\title{Plot the next most probable mutational events}
\usage{
plot_next_mutations(
  n,
  tree_df,
  Theta,
  mutations = NULL,
  tree_label = NULL,
  top_M = 5
)
}
\arguments{
\item{n}{Number of mutational events.}

\item{mutations}{A list of mutation names, which must be unique values.
If no names are given, then the mutation IDs will be used.}

\item{tree_label}{The title of the tree (Default: NULL).}

\item{top_M}{Number of most probable mutational events to plot (Default: 5).}

\item{tree}{A tree in named list format.}
}
\description{
Given a particular tree and a Mutual Hazard Network, this function 
finds the next most probable mutational events.
}
