% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fns.R
\name{plot_observed_pathways}
\alias{plot_observed_pathways}
\title{Plot most frequent observed pathways from a cohort of mutation trees}
\usage{
plot_observed_pathways(
  tree_obj,
  Theta,
  top_M = 10,
  lambda_s = 1,
  mutation_colors = NULL,
  prob_digits = 2
)
}
\arguments{
\item{tree_obj}{A TreeMHN object.}

\item{Theta}{An n-by-n matrix representing a Mutual Hazard Network.}

\item{top_M}{Number of most frequent pathways to plot (Default: 10).}

\item{lambda_s}{Sampling rate (Default: 1)}

\item{mutation_colors}{A named vector with the color codes for all mutations (Default: NULL)}

\item{prob_digits}{Number of digits to show for the probabilities (Default: 2)}
}
\description{
This function takes a cohort of mutation trees and the estimated
Mutual Hazard Network as input and plot the most frequent observed pathways.
}
\author{
Xiang Ge Luo
}
