% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fns.R
\name{plot_pathways}
\alias{plot_pathways}
\title{Plot most probable pathways of a given length computed from a Mutual Hazard Network}
\usage{
plot_pathways(Theta, mutations = NULL, n_order = 4, top_M = 10, log2 = TRUE)
}
\arguments{
\item{Theta}{An n-by-n matrix representing a Mutual Hazard Network}

\item{mutations}{A list of mutation names, which must be unique values.
If no names are given, then the mutation IDs will be used.}

\item{n_order}{The length of the pathways (Default: 4).}

\item{top_M}{Number of most probable pathways to plot (Default: 10).}

\item{log2}{A boolean flag indicating whether the x axis is scaled by log2.}
}
\description{
This function takes a Mutual Hazard Network as input and plot the most probable
mutational pathways of a given length.
}
\author{
Xiang Ge Luo
}
