% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fns.R
\name{plot_pathways_w_sampling}
\alias{plot_pathways_w_sampling}
\title{Plot most probable pathways before a sampling event computed from a Mutual Hazard Network}
\usage{
plot_pathways_w_sampling(
  Theta,
  mutations,
  top_M = 10,
  lambda_s = 1,
  log2 = TRUE
)
}
\arguments{
\item{Theta}{An n-by-n matrix representing a Mutual Hazard Network.}

\item{mutations}{A list of mutation names, which must be unique values.
If no names are given, then the mutation IDs will be used.}

\item{top_M}{Number of most probable pathways to plot (Default: 10).}

\item{lambda_s}{Sampling rate (Default: 1)}

\item{log2}{A boolean flag indicating whether the x axis is scaled by log2.}
}
\description{
This function takes a Mutual Hazard Network as input and plot the most probable
mutational pathways before a sampling event.
}
\author{
Xiang Ge Luo
}
