% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fns.R
\name{plot_Theta}
\alias{plot_Theta}
\title{Plot a Mutual Hazard Network}
\usage{
plot_Theta(
  Theta,
  mutations = NULL,
  full = TRUE,
  sort_diag = TRUE,
  to_show = NULL
)
}
\arguments{
\item{Theta}{An n-by-n matrix representing a Mutual Hazard Network.}

\item{mutations}{A list of mutation names, which must be unique values.
If no names are given, then the mutation IDs will be used.}

\item{full}{A boolean flag indicating whether the full MHN is shown (default: TRUE).
If not, then only mutations with non-zero off-diagonal entries will be shown.}

\item{sort_diag}{A boolean flag indicating whether the rows and columns of the MHN are
sorted based on the decreasing order of the diagonal entries (default: TRUE).}

\item{to_show}{A vector indicating which rows and columns of the MHN are shown (default: NULL).}
}
\description{
This function plots a Mutual Hazard Network by separating the 
diagonal entries from the off-diagonal entries.
}
