% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fns.R
\name{plot_tree_df}
\alias{plot_tree_df}
\title{Plot a tree in data frame format}
\usage{
plot_tree_df(tree_df, mutations, tree_label = NULL)
}
\arguments{
\item{mutations}{A list of mutation names corresponding to the mutation IDs.}

\item{tree_label}{The title of the tree (Default: NULL).}

\item{tree}{A tree in data frame format.}
}
\description{
This function plots a tree in data frame format.
}
\author{
Xiang Ge Luo
}
