% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fns.R
\name{plot_tree_list}
\alias{plot_tree_list}
\title{Plot a tree in named list format}
\usage{
plot_tree_list(tree, mutations, tree_label = NULL)
}
\arguments{
\item{tree}{A tree in named list format.}

\item{mutations}{A list of mutation names corresponding to the mutation IDs.}

\item{tree_label}{The title of the tree (Default: NULL).}
}
\description{
This function plots a tree in named list format.
}
\author{
Xiang Ge Luo
}
