% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{random_Theta}
\alias{random_Theta}
\title{random_Theta(n, sparsity, exclusive_ratio)
This function randomly generates an MHN with log-transformed parameters.}
\usage{
random_Theta(n, sparsity = 0.5, exclusive_ratio = 0.5)
}
\arguments{
\item{n}{Number of mutational events}

\item{sparsity}{Percentage of off diagonal elements with a value of zero in the MHN. (Default: 0.5)}

\item{exclusive_ratio}{Percentage of non-zero elements that are negative,
meaning that one mutation is inhibiting another mutation (Default: 0.9)}
}
\value{
An n-by-n matrix representing the MHN
}
\description{
random_Theta(n, sparsity, exclusive_ratio)
This function randomly generates an MHN with log-transformed parameters.
}
\references{
Schill et al. (2020). Modelling cancer progression using Mutual Hazard Networks. Bioinformatics, 36(1), 241–249.
}
\author{
Rudolf Schill, Xiang Ge Luo
}
