% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_output.R
\name{sort_one_tree_df}
\alias{sort_one_tree_df}
\title{tree_df_to_trees(one_tree_df)
A helper function that sorts the parent IDs and the node IDs such that they are in ascending order}
\usage{
sort_one_tree_df(one_tree_df)
}
\arguments{
\item{one_tree_df}{A tree data frame for a single patient/tumor}
}
\value{
A sorted tree data frame
}
\description{
tree_df_to_trees(one_tree_df)
A helper function that sorts the parent IDs and the node IDs such that they are in ascending order
}
