% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessment.R
\name{Theta_to_pathways}
\alias{Theta_to_pathways}
\title{Compute the pathway probabilities given a Mutual Hazard Network}
\usage{
Theta_to_pathways(Theta, n_order = 4, prob_only = TRUE)
}
\arguments{
\item{Theta}{An MHN represented by a square matrix.}

\item{n_order}{Length of the pathways (Default: 4).}

\item{prob_only}{A Boolean value that determines whether to output only the pathway probabilities
or the data frame containing also the pathways (Default: TRUE).}
}
\value{
Pathway probabilities
}
\description{
This function computes the pathway probabilities given a Mutual Hazard Network.
}
\author{
Xiang Ge Luo
}
