% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_output.R
\name{tree_df_to_trees}
\alias{tree_df_to_trees}
\title{tree_df_to_trees(n, tree_df)
This function processes a data frame of mutation trees
and output a list of processed mutation trees}
\usage{
tree_df_to_trees(n, tree_df)
}
\arguments{
\item{n}{Number of mutational events}

\item{tree_df}{A data frame with the following columns:
\itemize{
\item Tree_ID: IDs of mutation trees, unique for each patient
\item Node_ID: IDs of each node in the tree, including the root node (with ID "1"), unique for each node
\item Mutation_ID: IDs of each mutational event, the root node has a mutation ID of "0",
other mutation IDs can be duplicated in the tree to allow for parallel mutations
\item Parent_ID: IDs of the parent node ID. The root node has itself as parent (ID "1").
}}
}
\value{
A named list of a tree data frame and the corresponding trees
}
\description{
tree_df_to_trees(n, tree_df)
This function processes a data frame of mutation trees
and output a list of processed mutation trees
}
