% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn_MHN.R
\name{get_MC_flags}
\alias{get_MC_flags}
\title{get_MC_flags(N, trees)
A function that determines whether Monte Carlo sampling is needed for each tree}
\usage{
get_MC_flags(N, n, trees)
}
\arguments{
\item{N}{Sample size}

\item{n}{Number of events}

\item{trees}{Mutation tree structures}
}
\value{
A boolean vector indicating whether Monte Carlo sampling is used for each tree
}
\description{
get_MC_flags(N, trees)
A function that determines whether Monte Carlo sampling is needed for each tree
}
