% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn_MHN.R
\name{learn_MHN}
\alias{learn_MHN}
\title{Learn a Mutual Hazard Network from a set of mutation trees}
\usage{
learn_MHN(
  tree_obj,
  gamma = 0.5,
  lambda_s = 1,
  Theta_init = NULL,
  M = 100,
  iterations = 500,
  to_mask = integer(0),
  use_EM = FALSE,
  verbose = FALSE,
  MC_threshold = 500,
  increment_M = 0,
  increment_M_bound = 500,
  return_Theta_only = TRUE
)
}
\arguments{
\item{tree_obj}{A TreeMHN object}

\item{gamma}{Penalization parameter in the objective function (Default: 0.5).}

\item{lambda_s}{Sampling rate (Default: 1).}

\item{Theta_init}{Initial value of the MHN provided to the optimization procedure (Default: NULL).}

\item{M}{Number of Monte Carlo samples to be drawn (Default: 100).}

\item{iterations}{Number of iterations for the EM/MCEM algorithm (Default: 500).}

\item{to_mask}{An integer vector of indices by column, which is used to mask the
off-diagonal entries of an MHN (Default: an empty vector).}

\item{use_EM}{A boolean value to determine whether the EM/MCEM algorithm is used (Default: FALSE).}

\item{verbose}{A boolean value to determine whether optimization steps are printed (Default: FALSE).}

\item{MC_threshold}{A threshold on the maximum number of subtrees of a given tree,
above which Monte Carlo sampling will be used (Default: 500).}

\item{increment_M}{The step size to increment the number of Monte Carlo samples (Default: 0).}

\item{increment_M_bound}{The upper bound on the number of Monte Carlo samples (Default: 500).}

\item{A}{boolean value to determine whether the function returns only the estimated
Theta or the TreeMHN object containing the estimated Theta and other parameters (Default: TRUE).}
}
\value{
A Mutual Hazard Network Theta
}
\description{
This function learns a Mutual Hazard Network from a set of mutation trees
in the format of a TreeMHN object.
}
\author{
Xiang Ge Luo
}
