% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fns.R
\name{plot_observed_pathways}
\alias{plot_observed_pathways}
\title{Plot most frequent observed pathways from a cohort of mutation trees}
\usage{
plot_observed_pathways(tree_obj, Theta, top_M = 10, lambda_s = 1, log2 = TRUE)
}
\arguments{
\item{tree_obj}{A TreeMHN object.}

\item{Theta}{An n-by-n matrix representing a Mutual Hazard Network.}

\item{top_M}{Number of most frequent pathways to plot (Default: 10).}

\item{lambda_s}{Sampling rate (Default: 1)}

\item{log2}{A boolean flag indicating whether the x axis is scaled by log2.}
}
\description{
This function takes a cohort of mutation trees and the estimated
Mutual Hazard Network as input and plot the most frequent observed pathways.
}
\author{
Xiang Ge Luo
}
