% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fns.R
\name{plot_pathways_w_sampling}
\alias{plot_pathways_w_sampling}
\title{Plot most probable pathways before a sampling event computed from a Mutual Hazard Network}
\usage{
plot_pathways_w_sampling(
  Theta,
  mutations,
  top_M = 10,
  lambda_s = 1,
  mutation_colors = NULL,
  prob_digits = 2,
  full = FALSE
)
}
\arguments{
\item{Theta}{An n-by-n matrix representing a Mutual Hazard Network.}

\item{mutations}{A list of mutation names, which must be unique values.
If no names are given, then the mutation IDs will be used.}

\item{top_M}{Number of most probable pathways to plot (Default: 10).}

\item{lambda_s}{Sampling rate (Default: 1)}

\item{mutation_colors}{A named vector with the color codes for all mutations (Default: NULL)}

\item{prob_digits}{Number of digits to show for the probabilities (Default: 2)}

\item{full}{A boolean flag indicating whether the full MHN is used.
If false, then only mutations with non-zero off-diagonal entries will be used (default: FALSE).}
}
\description{
This function takes a Mutual Hazard Network as input and plot the most probable
mutational pathways before a sampling event.
}
\author{
Xiang Ge Luo
}
